requirements.txt
python-dotenv
openai
requests
backports.zoneinfo;python_version<"3.9"
tzdata